requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
faiss-cpu>=1.7.4
openai>=1.3.0
tiktoken>=0.5.1
python-dotenv>=1.0.0
streamlit>=1.30.0